Fix Spawning Logic

+++ code/controllers/subsystem/job.dm
@@ -556,12 +556,13 @@ SUBSYSTEM_DEF(job)
 			else
 				M = living_mob
 		else
+			if(!isnull(new_mob)) //Detect fail condition on equip
 			//if equip() is somehow able to fail, send them back to lobby
-			var/mob/dead/new_player/NP = new()
-			NP.ckey = M.client.ckey
-			qdel(M)
-			to_chat(M, "Error equipping [rank]. Returning to lobby.</b>")
-			return null
+				var/mob/dead/new_player/NP = new()
+				NP.ckey = M.client.ckey
+				qdel(M)
+				to_chat(M, "Error equipping [rank]. Returning to lobby.</b>")
+				return null
 		SSpersistence.antag_rep_change[M.client.ckey] += job.GetAntagRep()
 
 		if(M.client.holder)
